(Quasi) Allow more than one concurrent computation
The code is now adapted to allow as much as n concurrent computations where
n is the number of cores instead of only just one. However, I still have to
limit the number of concurrent compuations effectively to one (cannot compile
with -threaded flag) due to an unfortunate constellation of my server's kernel
version and ghc's version.